# Documentations/API_Documentation.yaml
openapi: 3.0.1
info:
  title: Game API
  description: API documentation for the game backend services.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /entity/heroes:
    get:
      summary: Get a list of heroes
      tags:
        - Entity
      responses:
        200:
          description: List of heroes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    type:
                      type: string
                    level:
                      type: integer
                    dunjonNb:
                      type: integer
                    id:
                      type: integer
                    maxHP:
                      type: integer
                    gold:
                      type: number
                    ATK:
                      type: number

    post:
      summary: Create a new hero
      tags:
        - Entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        201:
          description: Hero created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
  /entity/heroes/:
    put:
      summary: Update an existing hero
      tags:
        - Entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                level:
                  type: integer
                dunjonNb:
                  type: integer
                maxHP:
                  type: integer
                atk:
                  type: integer
                gold:
                  type: integer
              required:
                - name
                - type
                - level
                - dunjonNb
                - maxHP
                - atk
                - gold
      responses:
        200:
          description: Hero updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  type:
                    type: string
                  level:
                    type: integer
                  dunjonNb:
                    type: integer
                  maxHP:
                    type: integer
                  ATK:
                    type: integer
                  gold:
                    type: integer
...